12-27-2018/requirements.txt
nose
mock
PyHamcrest
coverage
flake8
parameterized
